Cardano.Wallet.Byron

Description

This module provides the main wallet server function for the Byron-rewrite Haskell node backend.

The Cardano.Wallet.Byron.Network uses the mini-protocols (ChainSync and TxSubmission) to talk with a core node and synchronize with the network.

Functionality specific to this backend for creating transactions is in Cardano.Wallet.Byron.Transaction

Documentation

serveWallet

Arguments

:: forall (n :: NetworkDiscriminant). (NetworkDiscriminantVal n, DecodeAddress n, EncodeAddress n, WorstSizeOf Address n IcarusKey, WorstSizeOf Address n ByronKey, t ~ IO Byron)

=> Tracers IO
Logging config.

-> SyncTolerance
A time tolerance within we consider being synced

-> Maybe FilePath
Database folder filepath

-> HostPreference
Which host to bind.

-> Listen
HTTP API Server port.

-> AddrInfo
Socket for communicating with the node

-> Block
The genesis block, or some starting point.

-> (BlockchainParameters, (NodeToClientVersionData, CodecCBORTerm Text NodeToClientVersionData))
Network parameters needed to connect to the underlying network.

See also: 'Cardano.Wallet.Byron.Compatibility#KnownNetwork'.

-> (SockAddr -> IO ())
Callback to run before the main loop

-> IO ExitCode

The cardano-wallet-shelley main function. It takes the configuration which was passed from the CLI and environment and starts all components of the wallet.

Tracing

data Tracers' f

The types of trace events produced by the Byron API server.

Constructors

Tracers

Fields

applicationTracer :: f ApplicationLog

apiServerTracer :: f ApiLog

walletEngineTracer :: f (WorkerLog WalletId WalletLog)

walletDbTracer :: f DBLog

ntpClientTracer :: f NtpTrace

type Tracers m = Tracers' (Tracer m)

All of the Byron Tracers.

type TracerSeverities = Tracers' (Const (Maybe Severity))

The minimum severities for Tracers'. Nothing indicates that tracing is completely disabled.

tracerLabels :: Tracers' (Const Text)

Strings that the user can refer to tracers by.

tracerDescriptions :: [(String, String)]

Names and descriptions of the tracers, for user documentation.

setupTracers :: TracerSeverities -> Trace IO Text -> Tracers IO

Set up tracing with textual log messages.

tracerSeverities :: Maybe Severity -> TracerSeverities

Construct a TracerSeverities record with all tracers set to the given severity.

Logs

data ApplicationLog

Log messages related to application startup and shutdown.

Constructors

MsgStarting AddrInfo

MsgNetworkName NetworkDiscriminant

MsgServerStartupError ListenError

MsgDatabaseStartup DatabasesStartupLog