Cardano.Wallet.Byron.Compatibility

Description

Conversion functions and static chain settings for Byron.

Documentation

data Byron

data ByronBlock

type NodeVersionData = (NodeToClientVersionData, CodecCBORTerm Text NodeToClientVersionData)

Chain Parameters

mainnetVersionData :: NodeVersionData

Settings for configuring a MainNet network client

testnetVersionData :: ProtocolMagic -> NodeVersionData

Settings for configuring a TestNet network client

mainnetBlockchainParameters :: GenesisBlockParameters

Genesis

emptyGenesis :: BlockchainParameters -> Block

genesisTip :: Tip ByronBlock

Conversions

toByronHash :: Hash "BlockHeader" -> ByronHash

toEpochSlots :: EpochLength -> EpochSlots

toGenTx :: HasCallStack => SealedTx -> GenTx ByronBlock

SealedTx are the result of rightfully constructed byron transactions so, it is relatively safe to unserialize them from CBOR.

toPoint :: Hash "Genesis" -> EpochLength -> BlockHeader -> Point ByronBlock

toSlotNo :: EpochLength -> SlotId -> SlotNo

fromBlockNo :: BlockNo -> Quantity "block" Word32

fromByronBlock :: Hash "Genesis" -> EpochLength -> ByronBlock -> Block

fromByronHash :: ByronHash -> Hash "BlockHeader"

fromChainHash :: Hash "Genesis" -> ChainHash ByronBlock -> Hash "BlockHeader"

fromGenesisData :: (GenesisData, GenesisHash) -> (GenesisBlockParameters, [TxOut])

Convert genesis data into blockchain params and an initial set of UTxO

fromNetworkMagic :: NetworkMagic -> ProtocolMagic

fromProtocolMagicId :: ProtocolMagicId -> ProtocolMagic

fromSlotNo :: EpochLength -> SlotNo -> SlotId

fromTip :: Hash "Genesis" -> EpochLength -> Tip ByronBlock -> BlockHeader

fromTxAux :: TxAux -> Tx

fromTxIn :: TxIn -> TxIn

fromTxOut :: TxOut -> TxOut

txParametersFromUpdateState :: State -> TxParameters

Pluck the blockchain parameters relevant to creating transactions out of the cardano-chain update state record.

Orphan instances

EncodeAddress Mainnet

Instance details

Methods

encodeAddress :: Address -> Text

DecodeAddress Mainnet

Instance details

Methods

decodeAddress :: Text -> Either TextDecodingError Address

EncodeAddress (Testnet pm)

Instance details

Methods

encodeAddress :: Address -> Text

KnownNat pm => DecodeAddress (Testnet pm)

Instance details

Methods

decodeAddress :: Text -> Either TextDecodingError Address