Cardano.Wallet.Byron.Compatibility

Description

Conversion functions and static chain settings for Byron.

Documentation

data Byron

data ByronBlock

type NodeVersionData = (NodeToClientVersionData, CodecCBORTerm Text NodeToClientVersionData)

Chain Parameters

class KnownNetwork (n :: NetworkDiscriminant) where

Embed some constants into a network type.

Methods

blockchainParameters :: BlockchainParameters

versionData :: NodeVersionData

mainnetGenesis :: Block

Genesis

genesisTip :: Tip ByronBlock

Conversions

toByronHash :: Hash "BlockHeader" -> ByronHash

toEpochSlots :: EpochLength -> EpochSlots

toGenTx :: HasCallStack => SealedTx -> GenTx ByronBlock

SealedTx are the result of rightfully constructed byron transactions so, it is relatively safe to unserialize them from CBOR.

toPoint :: EpochLength -> BlockHeader -> Point ByronBlock

toSlotNo :: EpochLength -> SlotId -> SlotNo

fromBlockNo :: BlockNo -> Quantity "block" Word32

fromByronBlock :: Hash "Genesis" -> EpochLength -> ByronBlock -> Block

fromByronHash :: ByronHash -> Hash "BlockHeader"

fromChainHash :: Hash "Genesis" -> ChainHash ByronBlock -> Hash "BlockHeader"

fromGenesisData :: (GenesisData, GenesisHash) -> (BlockchainParameters, [TxOut])

Convert genesis data into blockchain params and an initial set of UTxO

fromSlotNo :: EpochLength -> SlotNo -> SlotId

fromTip :: Hash "Genesis" -> EpochLength -> Tip ByronBlock -> BlockHeader

fromTxAux :: TxAux -> Tx

fromTxIn :: TxIn -> TxIn

fromTxOut :: TxOut -> TxOut